backout: Backed out fix for gh-6 and gh-7.
The fix was correct for ZeroMQ 3.2 or higher. So I decided to rollback the change and use another branch for an nzmqt version compatible with newer ZeroMQ versions.

+++ include/nzmqt/nzmqt.hpp
@@ -361,7 +361,7 @@ namespace nzmqt
         // to be received.
         inline bool hasMoreMessageParts() const
         {
-            int value;
+            qint64 value;
             size_t size = sizeof(value);
             getOption(OPT_RCVMORE, &value, &size);
             return value;
